+++ src/main/java/duong/thuy/parking/services/CredentialsService.java
@@ -11,5 +11,5 @@ public interface CredentialsService {
 
     ResponseEntity<ResponseData> login(LoginRequest request);
 
-    ResponseEntity<ResponseData> getUserProfile(int userId);
+    ResponseEntity<ResponseData> getUserProfile(Long userId);
 }
