+++ src/main/java/duong/thuy/parking/services/FmcService.java
@@ -5,5 +5,5 @@ import duong.thuy.parking.response.ResponseData;
 import org.springframework.http.ResponseEntity;
 
 public interface FmcService {
-    ResponseEntity<ResponseData> requestUpdateFCMToken(FmcRequest request, int userId);
+    ResponseEntity<ResponseData> requestUpdateFCMToken(FmcRequest request, Long userId);
 }
