+++ src/main/java/duong/thuy/parking/services/OwnerService.java
@@ -5,5 +5,5 @@ import duong.thuy.parking.response.ResponseData;
 import org.springframework.http.ResponseEntity;
 
 public interface OwnerService {
-    public ResponseEntity<ResponseData> createOwner(CreateOwnerRequest request, int userId);
+    public ResponseEntity<ResponseData> createOwner(CreateOwnerRequest request, Long userId);
 }
